code improvement creation of Database.py and parsing of assignees partial
